+++ ProesBack/Infrastructure/Data/Migrations/ProesContextModelSnapshot.cs
@@ -24,11 +24,11 @@ namespace ProesBack.Infrastructure.Data.Migrations
 
             modelBuilder.Entity("ProesBack.Domain.Entities.Login", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint");
+                        .HasColumnType("int");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
                     b.Property<string>("Password")
                         .IsRequired()
@@ -56,11 +56,11 @@ namespace ProesBack.Infrastructure.Data.Migrations
 
             modelBuilder.Entity("ProesBack.Domain.Entities.Notification", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint");
+                        .HasColumnType("int");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
                     b.Property<DateTime>("Date")
                         .HasColumnType("datetime2");
@@ -86,11 +86,11 @@ namespace ProesBack.Infrastructure.Data.Migrations
 
             modelBuilder.Entity("ProesBack.Domain.Entities.User", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint");
+                        .HasColumnType("int");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
                     b.Property<DateTime>("BirthDate")
                         .HasColumnType("datetime2");
@@ -109,8 +109,8 @@ namespace ProesBack.Infrastructure.Data.Migrations
                         .HasMaxLength(200)
                         .HasColumnType("nvarchar(200)");
 
-                    b.Property<long>("loginId")
-                        .HasColumnType("bigint");
+                    b.Property<int>("loginId")
+                        .HasColumnType("int");
 
                     b.HasKey("Id");
